scripts/training/selim/entry_extraction/scripts/requirements.txt
transformers==4.8.2
torchmetrics==0.4.1
pytorch-lightning==1.3.8
tqdm==4.41.1
mlflow==1.19.0
scikit-learn==0.22.2.post1
sagemaker==2.49.1
s3fs==2021.07.0
smdebug==1.0.11
pytest==7.0